Widen scenario time fields to float and add timing.py constants module

+++ server/session.py
@@ -21,16 +21,16 @@ from typing import Optional
 @dataclass
 class Session:
     status: str  # preinit | ready | running_auto | running_manual | completed
-    simulation_time: int
+    simulation_time: float
     start_wall_time: Optional[float]
     events: list[dict]
     next_event_index: int
 
     environment: str
-    duration: int
+    duration: float
     baseline_metrics: dict
     eval_sql: str = ""
-    condition_at: Optional[int] = None
+    condition_at: Optional[float] = None
 
     # MicroMail
     micromail_emails: list[dict] = field(default_factory=list)
